maintenance: move src/auth.ts back to src/auth/auth.ts to help with merge
