+++ src/auth/auth.ts
@@ -1,13 +1,13 @@
 import FilenSDK, { APIError, FilenSDKConfig } from "@filen/sdk"
-import { err, errExit, out, outVerbose, prompt, promptConfirm } from "./interface/interface"
+import { err, errExit, out, outVerbose, prompt, promptConfirm } from "../interface/interface"
 import fs from "node:fs"
-import { exists, platformConfigPath } from "./util/util"
+import { exists, platformConfigPath } from "../util/util"
 import path from "path"
-import { wrapRedTerminalText } from "./interface/util"
-import { ANONYMOUS_SDK_CONFIG } from "./constants"
+import { wrapRedTerminalText } from "../interface/util"
+import { ANONYMOUS_SDK_CONFIG } from "../constants"
 import keytar from "keytar"
 import crypto from "node:crypto"
-import { isDevelopment } from "./index"
+import { isDevelopment } from "../index"
 
 /**
  * Handles authentication.
