+++ src/index.ts
@@ -3,7 +3,7 @@ import FilenSDK from "@filen/sdk"
 import path from "path"
 import os from "os"
 import { err, errExit, out, outVerbose, setOutputFlags, setupLogs } from "./interface/interface"
-import { Authentication } from "./auth"
+import { Authentication } from "./auth/auth"
 import { checkInjectedBuildInfo, version } from "./buildInfo"
 import { Updater } from "./updater"
 import { HelpPage } from "./interface/helpPage"
